Move updateTeamNotification from VoicesRepositoryImpl to NotificationsRepositoryImpl
