+++ app/src/test/java/org/ole/planet/myplanet/repository/VoicesRepositoryNewsSyncTest.kt
@@ -76,7 +76,6 @@ class VoicesRepositoryNewsSyncTest {
             Gson(),
             Gson(),
             sharedPrefManager,
-            mockk<TeamNotificationDao>(relaxed = true),
             newsDao,
             mockk<MyLibraryDao>(relaxed = true),
             mockk<org.ole.planet.myplanet.data.room.dao.NewsLogDao>(relaxed = true)
