+++ app/src/test/java/org/ole/planet/myplanet/ui/teams/voices/TeamsVoicesViewModelTest.kt
@@ -29,11 +29,12 @@ class TeamsVoicesViewModelTest {
     private val teamsRepository: TeamsRepository = mockk(relaxed = true)
     private val userRepository: org.ole.planet.myplanet.repository.UserRepository = mockk(relaxed = true)
     private val resourcesRepository: org.ole.planet.myplanet.repository.ResourcesRepository = mockk(relaxed = true)
+    private val notificationsRepository: org.ole.planet.myplanet.repository.NotificationsRepository = mockk(relaxed = true)
     private val dispatcherProvider = TestDispatcherProvider(testDispatcher)
 
     @Before
     fun setup() {
-        viewModel = TeamsVoicesViewModel(voicesRepository, teamsRepository, dispatcherProvider, userRepository, resourcesRepository)
+        viewModel = TeamsVoicesViewModel(voicesRepository, teamsRepository, dispatcherProvider, userRepository, resourcesRepository, notificationsRepository)
     }
 
     @Test
